requirements.txt
discord.py
python-dotenv
aiohttp
aiofiles
aiosqlite